refactor: 상품 생성 조건 도메인으로 이동
