+++ src/main/java/kitchenpos/application/ProductService.java
@@ -2,9 +2,7 @@ package kitchenpos.application;
 
 import static java.util.stream.Collectors.toList;
 
-import java.math.BigDecimal;
 import java.util.List;
-import java.util.Objects;
 import kitchenpos.application.request.ProductRequest;
 import kitchenpos.application.response.ProductResponse;
 import kitchenpos.dao.ProductDao;
@@ -23,13 +21,7 @@ public class ProductService {
 
     @Transactional
     public ProductResponse create(final ProductRequest request) {
-        final Product product = new Product(request.getName(), request.getPrice());
-        final BigDecimal price = product.getPrice();
-
-        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
-            throw new IllegalArgumentException("올바르지 않은 상품 가격입니다.");
-        }
-
+        final Product product = Product.of(request.getName(), request.getPrice());
         final Product savedProduct = productDao.save(product);
 
         return new ProductResponse(savedProduct);
