+++ src/main/java/kitchenpos/domain/Product.java
@@ -1,6 +1,7 @@
 package kitchenpos.domain;
 
 import java.math.BigDecimal;
+import java.util.Objects;
 
 public class Product {
 
@@ -11,9 +12,17 @@ public class Product {
     public Product() {
     }
 
-    public Product(final String name, final BigDecimal price) {
+    private Product(final String name, final BigDecimal price) {
         this.name = name;
         this.price = price;
+    }
+
+    public static Product of(final String name, final BigDecimal price) {
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException("올바르지 않은 상품 가격입니다.");
+        }
+
+        return new Product(name, price);
     }
 
     public Long getId() {
